+++ src/Stub/Factory/StubVerifierFactory.php
@@ -30,7 +30,7 @@ use Eloquent\Phony\Stub\StubInterface;
 use Eloquent\Phony\Stub\StubVerifier;
 
 /**
- * Creates stub verifierss.
+ * Creates stub verifiers.
  *
  * @internal
  */
@@ -219,20 +219,22 @@ class StubVerifierFactory implements StubVerifierFactoryInterface
     /**
      * Create a new stub verifier for the supplied callback.
      *
-     * @param callable|null $callback            The callback, or null to create an unbound stub verifier.
-     * @param mixed         $self                The self value.
-     * @param boolean|null  $useGeneratorSpies   True if generator spies should be used.
-     * @param boolean|null  $useTraversableSpies True if traversable spies should be used.
+     * @param callable|null $callback              The callback, or null to create an unbound stub verifier.
+     * @param mixed         $self                  The self value.
+     * @param callable|null $defaultAnswerCallback The callback to use when creating a default answer.
+     * @param boolean|null  $useGeneratorSpies     True if generator spies should be used.
+     * @param boolean|null  $useTraversableSpies   True if traversable spies should be used.
      *
      * @return StubVerifierInterface The newly created stub verifier.
      */
     public function createFromCallback(
         $callback = null,
         $self = null,
+        $defaultAnswerCallback = null,
         $useGeneratorSpies = null,
         $useTraversableSpies = null
     ) {
-        $stub = $this->stubFactory->create($callback, $self);
+        $stub = $this->stubFactory->create($callback, $self, $defaultAnswerCallback);
 
         return $this->create(
             $stub,