+++ src/Stub/Factory/StubVerifierFactoryInterface.php
@@ -36,16 +36,18 @@ interface StubVerifierFactoryInterface
     /**
      * Create a new stub verifier for the supplied callback.
      *
-     * @param callable|null $callback            The callback, or null to create an unbound stub verifier.
-     * @param mixed         $self                The self value.
-     * @param boolean|null  $useGeneratorSpies   True if generator spies should be used.
-     * @param boolean|null  $useTraversableSpies True if traversable spies should be used.
+     * @param callable|null $callback              The callback, or null to create an unbound stub verifier.
+     * @param mixed         $self                  The self value.
+     * @param callable|null $defaultAnswerCallback The callback to use when creating a default answer.
+     * @param boolean|null  $useGeneratorSpies     True if generator spies should be used.
+     * @param boolean|null  $useTraversableSpies   True if traversable spies should be used.
      *
      * @return StubVerifierInterface The newly created stub verifier.
      */
     public function createFromCallback(
         $callback = null,
         $self = null,
+        $defaultAnswerCallback = null,
         $useGeneratorSpies = null,
         $useTraversableSpies = null
     );